backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==9.5.0
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
